I converted the health system to float. I also changed various things to round the display as well. The regen system now supports regen rates > 20 for aliens.

+++ trunk/src/game/g_local.h
@@ -162,7 +162,7 @@ struct gentity_s
   int               fly_sound_debounce_time;  // wind tunnel
   int               last_move_time;
 
-  int               health;
+  float             health;
   int               lastHealth; // currently only used for overmind
 
   qboolean          takedamage;
